BSpline coords to float; RGB support

+++ src/plugins/opencv/bspline.h
@@ -11,25 +11,25 @@ template <unsigned DEGREE>
 class BSpline {
   public:
 	BSpline(unsigned subdiv = 0,
-	        const std::array<double, DEGREE>& min = initArray(0.0),
-	        const std::array<double, DEGREE>& max = initArray(1.0));
+	        const std::array<float, DEGREE>& min = initArray(0.0),
+	        const std::array<float, DEGREE>& max = initArray(1.0));
 
-	void addSample(const std::array<double, DEGREE>& coords, double value);
-	double sample(const std::array<double, DEGREE>& coords) const;
+	void addSample(const std::array<float, DEGREE>& coords, float value);
+	float sample(const std::array<float, DEGREE>& coords) const;
 
 	bool operator==(const BSpline& b) const;
 	bool operator!=(const BSpline& b) const;
 
   private:
-	static double B(double t, unsigned k);
+	static float B(float t, unsigned k);
 
 	template <typename FN>
-	inline void visit(const std::array<double, DEGREE>& coords, const FN& fn) const;
-	static std::array<double, DEGREE> initArray(double val);
+	inline void visit(const std::array<float, DEGREE>& coords, const FN& fn) const;
+	static std::array<float, DEGREE> initArray(float val);
 
 	std::size_t m_subdiv;
 	std::vector<std::pair<float, float>> m_controls;
-	std::array<double, DEGREE> m_min, m_max;
+	std::array<float, DEGREE> m_min, m_max;
 };
 
 template <unsigned DEGREE>
